Revert "revert"
This reverts commit f3e0b5e9d8b382c872677d683619ef94f47914e2.

+++ src/entities/Account.ts
@@ -27,7 +27,7 @@ export class Account extends BasicEntity {
   })
   password!: string
 
-  @Column("decimal", {
+  @Column("int", {
     unique: true,
   })
   accountNumber!: number
